changing to maximal downsampling and padding approach

+++ DiffTransformerV3/dataset.py
@@ -71,7 +71,7 @@ class TensorStackDataset(Dataset):
             feature_stack = torch.from_numpy(feature_stack)
 
             # make sure that the feature stack is Double
-        feature_stack = feature_stack.double()
+        feature_stack = feature_stack.float()
 
         diff_list = []
